+++ core/CHANGES.rst
@@ -3,6 +3,26 @@ vivi.core changes
 
 .. towncrier release notes start
 
+6.75.0 (2024-11-22)
+-------------------
+
+- WCM-551: Update body cache when loading content anyway (prevents an extra sql query for each load)
+
+
+6.74.0 (2024-11-22)
+-------------------
+
+- WCM-551: Cache analyzed toggles, not just their XML nodes (5% performance gain)
+
+
+6.73.0 (2024-11-22)
+-------------------
+
+- WCM-544: read centerpage type from properties and not from xml body
+
+- WCM-466: remove free/dynamic access toggle
+
+
 6.72.0 (2024-11-20)
 -------------------